Split some functions from ui_utils to separate files
