+++ web/ui_utils/WaitOnType.js
@@ -1,0 +1,4 @@
+export default Object.freeze({
+  EVENT: 'event',
+  TIMEOUT: 'timeout',
+});
