+++ web/ui_utils/clamp.js
@@ -1,0 +1,3 @@
+export default function clamp(v, min, max) {
+  return Math.min(Math.max(v, min), max);
+}
