+++ web/ui_utils/getGlobalEventBus.js
@@ -1,0 +1,9 @@
+import EventBus from './EventBus';
+
+let globalEventBus = null;
+export default function getGlobalEventBus(dispatchToDOM = false) {
+  if (!globalEventBus) {
+    globalEventBus = new EventBus({ dispatchToDOM, });
+  }
+  return globalEventBus;
+}
