+++ web/ui_utils/isValidRotation.js
@@ -1,0 +1,3 @@
+export default function isValidRotation(angle) {
+  return Number.isInteger(angle) && angle % 90 === 0;
+}
